Expand selection range to match suggestions
This fixes some inconsistent behavior. Previously, if you selected a
range that included part of a suggestion at the end, and that suggestion
had another suggestion afterwards which was merged into it, clicking
accept/reject would not affect the merged in suggestion that's outside
the selection. However, clicking accept with a cursor selection (just
one pos) would resolve all merged suggestions.